+++ plugins/enpro26/router/tasks.py
@@ -12,28 +12,28 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import traceback
 import requests
 from pathlib import PurePath
-from typing import Optional
-from urllib.parse import urljoin
 from celery.utils.log import get_task_logger
 
 from qhana_plugin_runner.celery import CELERY
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
 from qhana_plugin_runner.plugin_utils.attributes import AttributeMetadata
-from qhana_plugin_runner.plugin_utils.entity_marshalling import (
-    EntityTupleMixin,
-    load_entities,
-)
-from qhana_plugin_runner.plugin_utils.interop import get_plugin_endpoint
+from qhana_plugin_runner.plugin_utils.entity_marshalling import load_entities
 from qhana_plugin_runner.plugin_utils.zip_utils import get_files_from_zip_url
 from qhana_plugin_runner.requests import get_mimetype, open_url
-from qhana_plugin_runner.storage import STORE
-from qhana_plugin_runner.tasks import save_task_error, save_task_result
 
 from . import Router
 from .schemas import InputParameters, InputParametersSchema
+from .task_helpers import _load_task, _load_entity_attributes, _taxonomy_ref
+from .pipeline_tasks import (
+    CELERY_COUNTDOWN,
+    launch_next_pipeline,
+    start_transformers,
+    start_aggregator,
+    start_mds,
+    finalize_pipeline,
+)
 
 TASK_LOGGER = get_task_logger(__name__)
 
@@ -49,83 +49,6 @@ PIPELINE_PLUGINS = {
     "aggregator": "attribute-distance-aggregator",
     "mds": "attribute-distance-mds",
 }
-
-
-# --- HELPER FUNCTIONS ---
-def subscribe_to_plugin(task_result_url: str, webhook_url: str):
-    """Subscribes the webhook to a target plugin's task result updates."""
-
-    webhook_url = webhook_url.replace("localhost", "127.0.0.1")
-    response = requests.get(task_result_url).json()
-    subscription_link = next(
-        (
-            link["href"]
-            for link in response.get("links", [])
-            if link["type"] == "subscribe"
-        ),
-        None,
-    )
-    if not subscription_link:
-        raise ValueError("Target plugin does not support subscriptions!")
-    requests.post(
-        subscription_link,
-        json={"command": "subscribe", "event": "status", "webhookHref": webhook_url},
-    ).raise_for_status()
-
-
-def extract_output_url(outputs: list, data_type: str) -> str:
-    """Finds the URL of a specific output from a plugin result based on its dataType."""
-    for output in outputs:
-        if output.get("dataType") == data_type:
-            return output["href"]
-    raise ValueError(f"Could not find output with dataType {data_type}")
-
-
-def _plugin_process_url(task_data: ProcessingTask, plugin: str) -> str:
-    """Resolve a pipeline plugin's process endpoint from its metadata url.
-
-    The route handler stores the external metadata url per plugin in
-    ``data["plugin_urls"]``. ``get_plugin_endpoint`` follows the metadata to
-    the plugin's processing endpoint.
-    """
-    return get_plugin_endpoint(task_data.data["plugin_urls"][plugin])
-
-
-def _load_task(db_id: int) -> ProcessingTask:
-    task_data: Optional[ProcessingTask] = ProcessingTask.get_by_id(id_=db_id)
-    if task_data is None:
-        msg = f"Could not load task data with id {db_id} to read parameters!"
-        TASK_LOGGER.error(msg)
-        raise KeyError(msg)
-    return task_data
-
-
-def _taxonomy_ref(metadata: AttributeMetadata) -> Optional[str]:
-    """Return the referenced taxonomy file name, or ``None`` for non-taxonomy attributes.
-
-    Taxonomy references point into the taxonomies zip, e.g.
-    ``taxonomies.zip:t_Gattung.json`` (the file name is the part after the
-    colon). Other references target plain files such as ``people.csv`` and are
-    excluded. The ``refTarget`` field is the reliable signal here. The ``type``
-    field carries the attribute id in the sample data, not the data type.
-    """
-    ref_target = metadata.ref_target or ""
-    if ".zip:" not in ref_target:
-        return None
-    return ref_target.split(":", 1)[1]
-
-
-def _load_entity_attributes(entities_url: str) -> set:
-    """Return the set of attribute names present in the entities file."""
-    attributes: set = set()
-    with open_url(entities_url) as response:
-        mimetype = get_mimetype(response)
-        for ent in load_entities(response, mimetype):
-            if isinstance(ent, EntityTupleMixin):
-                attributes.update(type(ent).entity_attributes)
-                break
-            attributes.update(ent.keys())
-    return attributes
 
 
 # --- Step 1: discover taxonomy attributes for the routing step ---
@@ -175,19 +98,7 @@ def preprocessing_task(self, db_id: int) -> str:
 @CELERY.task(name=f"{Router.instance.identifier}.route_task", bind=True)
 def route_task(self, db_id: int) -> str:
     task_data = ProcessingTask.get_by_id(id_=db_id)
-    if not task_data:
-        raise KeyError(f"Could not load task data with id {db_id}")
-    
-    # Evaluate attributes stored from the routing UI to build the queue
-    queue = []
-    if task_data.data.get("wu_palmer_attributes"):
-        queue.append("wu_palmer")
-    if task_data.data.get("mapping_attributes"):
-        queue.append("mapping")
-
-    task_data.data["pipeline_queue"] = queue
-    task_data.save(commit=True)
-
+   
     # Trigger the first pipeline
     launch_next_pipeline(task_data)
     return "Routing task started and pipeline queued."
@@ -203,7 +114,6 @@ def handle_webhook_task(self, db_id: int, source_url: str):
     
     current_pipeline = task_data.data.get("current_pipeline")
 
-    # Delegate to the specific pipeline's webhook handler
     if current_pipeline == "wu_palmer":
         handle_wu_palmer_progression(task_data, db_id, source_url)
         
@@ -215,292 +125,25 @@ def handle_webhook_task(self, db_id: int, source_url: str):
         return "Unrecognized pipeline state"
 
 
-def launch_next_pipeline(task_data: ProcessingTask):
-    """Pops the next pipeline from the queue and triggers it."""
-    queue = task_data.data.get("pipeline_queue", [])
-    
-    if not queue:
-        # All pipelines have been finished.
-        # TODO: Add Dimension reduction and vector merge here
-        save_task_result.delay("All Pipelines Completed Successfully!", task_data.id)
-        return
-
-    # Pop the next pipeline and update state
-    next_pipeline = queue.pop(0)
-    task_data.data["pipeline_queue"] = queue
-    task_data.data["current_pipeline"] = next_pipeline
-
-    # Route to the correct starting step
-    if next_pipeline == "wu_palmer":
-        task_data.add_task_log_entry("Starting Wu-Palmer Pipeline. Includes: Wu-Palmer, Transformer, Aggregator, MDS")
-        task_data.save(commit=True)
-        start_wu_palmer.apply_async(args=[task_data.id])
-    elif next_pipeline == "mapping":
-        task_data.add_task_log_entry("Starting Distance Mapping Pipeline. Includes: Distance Mapping, Aggregator, MDS")
-        task_data.save(commit=True)
-        start_mapping.apply_async(args=[task_data.id])
-
 def handle_wu_palmer_progression(task_data: ProcessingTask, db_id: int, source_url:str):
-    # Launch the dedicated Celery task for whichever step just finished!
+    """Handle progression of the wu-palmer pipeline"""
+
     if source_url == task_data.data.get("wu_palmer_url"):
-        start_transformers.apply_async(args=[db_id, source_url], countdown=4)
+        start_transformers.apply_async(args=[db_id, source_url], countdown=CELERY_COUNTDOWN)
     elif source_url == task_data.data.get("transformers_url"):
-        start_aggregator.apply_async(args=[db_id, source_url], countdown=4)
+        start_aggregator.apply_async(args=[db_id, source_url], countdown=CELERY_COUNTDOWN)
     elif source_url == task_data.data.get("aggregators_url"):
-        start_mds.apply_async(args=[db_id, source_url], countdown=4)
+        start_mds.apply_async(args=[db_id, source_url], countdown=CELERY_COUNTDOWN)
     elif source_url == task_data.data.get("mds_url"):
-        finalize_pipeline.apply_async(args=[db_id, source_url], countdown=4)
+        finalize_pipeline.apply_async(args=[db_id, source_url], countdown=CELERY_COUNTDOWN)
 
 def handle_mapping_progression(task_data: ProcessingTask, db_id: int, source_url:str):
+    """Handle progression of the mapping pipeline"""
+
     if source_url == task_data.data.get("mapping_url"):
-        start_aggregator.apply_async(args=[db_id, source_url], countdown=4)
+        start_aggregator.apply_async(args=[db_id, source_url], countdown=CELERY_COUNTDOWN)
     elif source_url == task_data.data.get("aggregators_url"):
-        start_mds.apply_async(args=[db_id, source_url], countdown=4)
+        start_mds.apply_async(args=[db_id, source_url], countdown=CELERY_COUNTDOWN)
     elif source_url == task_data.data.get("mds_url"):
-        finalize_pipeline.apply_async(args=[db_id, source_url], countdown=4)
+        finalize_pipeline.apply_async(args=[db_id, source_url], countdown=CELERY_COUNTDOWN)
 
-@CELERY.task(name=f"{Router.instance.identifier}.start_wu_palmer", bind=True, max_retries=10)
-def start_wu_palmer(self, db_id: int):
-    TASK_LOGGER.info("Starting Wu-Palmer Plugin")
-    task_data = ProcessingTask.get_by_id(db_id)
-    try:
-        params: InputParameters = InputParametersSchema().loads(task_data.parameters)
-        payload = {
-            "entitiesUrl": params.entities_url,
-            "entitiesMetadataUrl": params.entities_metadata_url,
-            "taxonomiesZipUrl": params.taxonomies_zip_url,
-            "attributes": task_data.data["wu_palmer_attributes"],
-            "root_is_part_of_hierarchy": str(params.root_is_part_of_hierarchy).lower(),
-        }
-
-        wu_palmer_url = _plugin_process_url(task_data, "wu_palmer")
-        response = requests.post(wu_palmer_url, data=payload, allow_redirects=False)
-        response.raise_for_status()
-
-        task_url = urljoin(wu_palmer_url, response.headers["Location"])
-        task_data.data["wu_palmer_url"] = task_url
-        task_data.add_task_log_entry("Started Wu-Palmer.")
-        task_data.save(commit=True)
-
-        subscribe_to_plugin(task_url, task_data.data["webhook_url"])
-    except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectionError) as e:
-        task_data.add_task_log_entry(f"Error occured during wu-palmer plugin. Attempting to retry:\n{e}")
-        task_data.save(commit=True)
-        raise self.retry(exc=e, countdown=3)
-    except Exception as e:
-        task_data.add_task_log_entry(f"CRASH in wu_palmer step:\n{traceback.format_exc()}")
-        save_task_error.delay(failing_task_id=self.request.id, db_id=db_id)
-
-@CELERY.task(name=f"{Router.instance.identifier}.start_mapping", bind=True, max_retries=10)
-def start_mapping(self, db_id: int):
-    TASK_LOGGER.info("Starting Mapping Plugin")
-    task_data = ProcessingTask.get_by_id(db_id)
-    try:
-        params: InputParameters = InputParametersSchema().loads(task_data.parameters)
-        payload = {
-            "entitiesUrl": params.entities_url,
-            "entitiesMetadataUrl": params.entities_metadata_url,
-            "taxonomiesZipUrl": params.taxonomies_zip_url,
-            "attributes": task_data.data["mapping_attributes"],
-            "distanceMetric": params.distance_metric.name,
-        }
-
-        mapping_url = _plugin_process_url(task_data, "numerical_mapping")
-        response = requests.post(mapping_url, data=payload, allow_redirects=False)
-        response.raise_for_status()
-
-        task_url = urljoin(mapping_url, response.headers["Location"])
-        task_data.data["mapping_url"] = task_url
-        task_data.add_task_log_entry("Started Mapping Distances.")
-        task_data.save(commit=True)
-        
-        subscribe_to_plugin(task_url, task_data.data["webhook_url"])
-    except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectionError) as e:
-        task_data.add_task_log_entry(f"Error occured during mapping plugin. Attempting to retry:\n{e}")
-        task_data.save(commit=True)
-        raise self.retry(exc=e, countdown=3)
-    except Exception as e:
-        task_data.add_task_log_entry(f"CRASH in mapping step:\n{traceback.format_exc()}")
-        save_task_error.delay(failing_task_id=self.request.id, db_id=db_id)
-
-
-# --- TRANSFORMER TASK ---
-@CELERY.task(
-    name=f"{Router.instance.identifier}.start_transformers",
-    bind=True,
-    max_retries=10,
-)
-def start_transformers(self, db_id: int, source_url: str):
-    TASK_LOGGER.info("Starting Transformers Plugin")
-    task_data = ProcessingTask.get_by_id(db_id)
-    try:
-        # 1. Persist Transformers result
-        outputs = requests.get(source_url).json().get("outputs", [])
-        element_sims_url = extract_output_url(outputs, "relation/element-similarities")
-        params = InputParametersSchema().loads(task_data.parameters)
-
-        if params.include_intermediate_results_in_output:
-            # Wu-Palmer result output
-            STORE.persist_task_result(
-                db_id,
-                open_url(element_sims_url).content,
-                "wu_palmer_similarities.zip",
-                "relation/element-similarities",
-                "application/zip",
-            )
-
-        # 2. Launch Transformers
-        payload = {
-            "similaritiesUrl": element_sims_url,
-            "attributes": task_data.data["wu_palmer_attributes"],
-            "transformer": params.transformer.name,
-        }
-
-        transformers_url = _plugin_process_url(task_data, "transformers")
-        response = requests.post(transformers_url, data=payload, allow_redirects=False)
-        task_url = urljoin(transformers_url, response.headers["Location"])
-
-        task_data.data["transformers_url"] = task_url
-        task_data.add_task_log_entry("Started Transformers.")
-        task_data.save(commit=True)
-        subscribe_to_plugin(task_url, task_data.data["webhook_url"])
-
-    except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectionError) as e:
-        task_data.add_task_log_entry(f"Error occured during transformers plugin. Attempting to retry:\n{e}")
-        task_data.save(commit=True)
-        raise self.retry(exc=e, countdown=3)
-    except Exception as e:
-        task_data.add_task_log_entry(f"CRASH in transformers step:\n{traceback.format_exc()}")
-        save_task_error.delay(failing_task_id=self.request.id, db_id=db_id)
-
-
-# --- AGGREGATOR TASK ---
-@CELERY.task(
-    name=f"{Router.instance.identifier}.start_aggregator",
-    bind=True,
-    max_retries=10,
-)
-def start_aggregator(self, db_id: int, source_url: str):
-    TASK_LOGGER.info("Starting Aggregator Plugin")
-    task_data = ProcessingTask.get_by_id(db_id)
-    try:
-        outputs = requests.get(source_url).json().get("outputs", [])
-        element_dists_url = extract_output_url(outputs, "relation/element-distances")
-        params = InputParametersSchema().loads(task_data.parameters)
-
-        if params.include_intermediate_results_in_output:
-            # Transformers result output
-            prefix = task_data.data.get("current_pipeline", "unknown")
-            STORE.persist_task_result(
-                db_id,
-                open_url(element_dists_url).content,
-                f"{prefix}_element_distances.zip",
-                "relation/element-distances",
-                "application/zip",
-            )
-
-        payload = {
-            "entitiesUrl": params.entities_url,
-            "elementDistancesUrl": element_dists_url,
-        }
-
-        agg_url = _plugin_process_url(task_data, "aggregator")
-        response = requests.post(agg_url, data=payload, allow_redirects=False)
-        task_url = urljoin(agg_url, response.headers["Location"])
-
-        task_data.data["aggregators_url"] = task_url
-        task_data.add_task_log_entry("Started Aggregator.")
-        task_data.save(commit=True)
-        subscribe_to_plugin(task_url, task_data.data["webhook_url"])
-
-    except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectionError) as e:
-        task_data.add_task_log_entry(f"Error occured during aggregator step. Attempting to retry:\n{e}")
-        task_data.save(commit=True)
-        raise self.retry(exc=e, countdown=3)
-    except Exception as e:
-        task_data.add_task_log_entry(f"CRASH in aggreagtor step:\n{traceback.format_exc()}")
-        save_task_error.delay(failing_task_id=self.request.id, db_id=db_id)
-
-
-# --- MDS TASK ---
-@CELERY.task(
-    name=f"{Router.instance.identifier}.start_mds", bind=True, max_retries=10
-)
-def start_mds(self, db_id: int, source_url: str):
-    TASK_LOGGER.info("Starting MDS Plugin")
-    task_data = ProcessingTask.get_by_id(db_id)
-    try:
-        outputs = requests.get(source_url).json().get("outputs", [])
-        attr_dists_url = extract_output_url(outputs, "relation/attribute-distances")
-        params = InputParametersSchema().loads(task_data.parameters)
-
-        if params.include_intermediate_results_in_output:
-            # Aggregator result output
-            prefix = task_data.data.get("current_pipeline", "unknown")
-            STORE.persist_task_result(
-                db_id,
-                open_url(attr_dists_url).content,
-                f"{prefix}_attribute_distances.zip",
-                "relation/attribute-distances",
-                "application/zip",
-            )
-
-        payload = {
-            "attributeDistancesUrl": attr_dists_url,
-            "dimensions": params.dimensions,
-            "metric": params.metric.name,
-            "nInit": params.n_init,
-            "maxIter": params.max_iter,
-            "missingDataHandling": params.missing_data_handling.name,
-        }
-
-        mds_url = _plugin_process_url(task_data, "mds")
-        response = requests.post(mds_url, data=payload, allow_redirects=False)
-
-        task_url = urljoin(mds_url, response.headers["Location"])
-        task_data.data["mds_url"] = task_url
-
-        task_data.add_task_log_entry("Started MDS.")
-        task_data.save(commit=True)
-
-        subscribe_to_plugin(task_url, task_data.data["webhook_url"])
-
-    except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectionError) as e:
-        task_data.add_task_log_entry(f"Error occured during mds step. Attempting to retry:\n{e}")
-        task_data.save(commit=True)
-        raise self.retry(exc=e, countdown=3)
-    except Exception as e:
-        task_data.add_task_log_entry(f"CRASH in mds step:\n{traceback.format_exc()}")
-        save_task_error.delay(failing_task_id=self.request.id, db_id=db_id)
-
-
-# --- END OF PIPELINE ---
-@CELERY.task(
-    name=f"{Router.instance.identifier}.finalize_pipeline", bind=True, max_retries=10
-)
-def finalize_pipeline(self, db_id: int, source_url: str):
-    TASK_LOGGER.info("Finishing the Pipeline")
-    task_data = ProcessingTask.get_by_id(db_id)
-    try:
-        outputs = requests.get(source_url).json().get("outputs", [])
-        final_dists_url = extract_output_url(outputs, "entity/vector")
-
-        prefix = task_data.data.get("current_pipeline", "unknown")
-        STORE.persist_task_result(
-            db_id,
-            open_url(final_dists_url).content,
-            f"{prefix}_mds_final_vectors.zip",
-            "entity/vector",
-            "application/zip",
-        )
-
-        # trigger next pipeline
-        launch_next_pipeline(task_data)
-
-    except (requests.exceptions.ReadTimeout, requests.exceptions.ConnectionError) as e:
-        task_data.add_task_log_entry(f"Error occured during finalization step. Attempting to retry:\n{e}")
-        task_data.save(commit=True)
-        raise self.retry(exc=e, countdown=3)
-    except Exception as e:
-        task_data.add_task_log_entry(f"CRASH in final step:\n{traceback.format_exc()}")
-        save_task_error.delay(failing_task_id=self.request.id, db_id=db_id)
